requirements.txt
colorama
numpy
scipy
pytorch
pytorch-transformers
